# manager/docs/v3/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: support@starton.io
    name: API Support
    url: https://starton.io
  description: This is the manager service for Tyrscale
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Tyrscale Manager API
  version: "1.0"
...
servers:
- url: /
paths:
...
  /plugins:
    get:
      description: Get list plugins
      operationId: listPlugins
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.DefaultSuccessResponse-plugin_ListPluginsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.BadRequestResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.InternalServerErrorResponse'
          description: Internal Server Error
      summary: Get list plugins
      tags:
      - plugins
...
  /routes:
    get:
      description: Get list routes
      operationId: listRoutes
      parameters:
      - in: query
        name: host
        schema:
          type: string
      - in: query
        name: loadBalancerStrategy
        schema:
          enum:
          - weight-round-robin
          - least-load
          - failover-priority
          type: string
          x-enum-varnames:
          - BalancerWeightRoundRobin
          - BalancerLeastLoad
          - BalancerPriority
        x-enum-varnames:
        - BalancerWeightRoundRobin
        - BalancerLeastLoad
        - BalancerPriority
      - in: query
        name: path
        schema:
          type: string
      - in: query
        name: uuid
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.DefaultSuccessResponse-ListRouteRes'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.BadRequestResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.InternalServerErrorResponse'
          description: Internal Server Error
      summary: Get list routes
      tags:
      - routes
    post:
      description: Create a route
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteReq'
        description: Route request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.CreatedSuccessResponse-CreateRouteRes'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.BadRequestResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.InternalServerErrorResponse'
          description: Internal Server Error
      summary: Create a route
      tags:
      - routes
      x-codegen-request-body-name: route
...
  /routes/{uuid}:
    delete:
      description: Delete a route
      operationId: deleteRoute
      parameters:
      - description: Route UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.DefaultSuccessResponseWithoutData'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.BadRequestResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.InternalServerErrorResponse'
          description: Internal Server Error
      summary: Delete a route
      tags:
      - routes
    put:
      description: Update a route
      operationId: updateRoute
      parameters:
      - description: UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRouteReq'
        description: Route request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.DefaultSuccessResponseWithoutData'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.BadRequestResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.InternalServerErrorResponse'
          description: Internal Server Error
      summary: Update a route
      tags:
      - routes
      x-codegen-request-body-name: route
  /routes/{uuid}/attach-plugin:
    post:
      description: Attach plugin to route
      operationId: attachPlugin
      parameters:
      - description: Route UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachPluginReq'
        description: Attach plugin request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.DefaultSuccessResponseWithoutData'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.BadRequestResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.InternalServerErrorResponse'
          description: Internal Server Error
      summary: Attach plugin to route
      tags:
      - plugins
      x-codegen-request-body-name: body
  /routes/{uuid}/detach-plugin:
    post:
      description: Detach plugin from route
      operationId: detachPlugin
      parameters:
      - description: Route UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetachPluginReq'
        description: Detach plugin request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.DefaultSuccessResponseWithoutData'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.BadRequestResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.InternalServerErrorResponse'
          description: Internal Server Error
      summary: Detach plugin from route
      tags:
      - plugins
      x-codegen-request-body-name: body
  /routes/{uuid}/plugins:
    get:
      description: Get list plugins from route
      operationId: listPluginsFromRoute
      parameters:
      - description: Route UUID
        in: path
        name: uuid
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.DefaultSuccessResponse-Plugins'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.BadRequestResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/responses.InternalServerErrorResponse'
          description: Internal Server Error
      summary: Get list plugins from route
      tags:
      - plugins
...
components:
  schemas:
    balancer.LoadBalancerStrategy:
      enum:
      - weight-round-robin
      - least-load
      - failover-priority
      type: string
      x-enum-varnames:
      - BalancerWeightRoundRobin
      - BalancerLeastLoad
      - BalancerPriority
    circuitbreaker.Settings:
      example:
        max_requests: 1
        max_consecutive_failures: 6
        name: name
        interval: 0
        enabled: true
        timeout: 5
      properties:
        enabled:
          type: boolean
        interval:
          type: integer
        max_consecutive_failures:
          type: integer
        max_requests:
          type: integer
        name:
          type: string
        timeout:
          type: integer
      type: object
...
    AttachPluginReq:
      example:
        name: name
        description: description
        priority: 81
        type: ResponseInterceptor
        config: "{}"
      properties:
        config:
          type: object
        description:
          type: string
        name:
          type: string
        priority:
          maximum: 1000
          minimum: 1
          type: integer
        type:
          $ref: '#/components/schemas/plugin.PluginType'
      required:
      - config
      - name
      - priority
      - type
      type: object
    DetachPluginReq:
      example:
        name: name
        type: ResponseInterceptor
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/plugin.PluginType'
      required:
      - name
      - type
      type: object
    Plugin:
      example:
        name: name
        description: description
        priority: 81
        config: "{}"
      properties:
        config:
          type: object
        description:
          type: string
        name:
          type: string
        priority:
          maximum: 1000
          minimum: 1
          type: integer
      required:
      - name
      - priority
      type: object
    Plugins:
      example:
        RequestInterceptor:
        - name: name
          description: description
          priority: 81
          config: "{}"
        - name: name
          description: description
          priority: 81
          config: "{}"
        Middleware:
        - name: name
          description: description
          priority: 81
          config: "{}"
        - name: name
          description: description
          priority: 81
          config: "{}"
        ResponseInterceptor:
        - name: name
          description: description
          priority: 81
          config: "{}"
        - name: name
          description: description
          priority: 81
          config: "{}"
      properties:
        Middleware:
          items:
            $ref: '#/components/schemas/Plugin'
          type: array
        RequestInterceptor:
          items:
            $ref: '#/components/schemas/Plugin'
          type: array
        ResponseInterceptor:
          items:
            $ref: '#/components/schemas/Plugin'
          type: array
      type: object
    CreateRouteReq:
      example:
        path: path
        host: host
        load_balancer_strategy: weight-round-robin
        circuit_breaker:
          max_requests: 1
          max_consecutive_failures: 6
          name: name
          interval: 0
          enabled: true
          timeout: 5
        health_check:
          request:
            headers:
              key: headers
            status_code: 2
            method: method
            body: body
          combined_with_circuit_breaker: true
          interval: 5
          type: eth_block_number
          enabled: true
          timeout: 7
        uuid: uuid
      properties:
        circuit_breaker:
          $ref: '#/components/schemas/circuitbreaker.Settings'
        health_check:
          $ref: '#/components/schemas/healthcheck.HealthCheckConfig'
        host:
          type: string
        load_balancer_strategy:
          $ref: '#/components/schemas/balancer.LoadBalancerStrategy'
        path:
          type: string
        uuid:
          type: string
      required:
      - host
      - load_balancer_strategy
      type: object
    CreateRouteRes:
      example:
        uuid: uuid
      properties:
        uuid:
          type: string
      type: object
    ListRouteRes:
      example:
        items:
        - path: path
          host: host
          load_balancer_strategy: weight-round-robin
          circuit_breaker:
            max_requests: 1
            max_consecutive_failures: 6
            name: name
            interval: 0
            enabled: true
            timeout: 5
          health_check:
            request:
              headers:
                key: headers
              status_code: 2
              method: method
              body: body
            combined_with_circuit_breaker: true
            interval: 5
            type: eth_block_number
            enabled: true
            timeout: 7
          uuid: uuid
        - path: path
          host: host
          load_balancer_strategy: weight-round-robin
          circuit_breaker:
            max_requests: 1
            max_consecutive_failures: 6
            name: name
            interval: 0
            enabled: true
            timeout: 5
          health_check:
            request:
              headers:
                key: headers
              status_code: 2
              method: method
              body: body
            combined_with_circuit_breaker: true
            interval: 5
            type: eth_block_number
            enabled: true
            timeout: 7
          uuid: uuid
      properties:
        items:
          items:
            $ref: '#/components/schemas/Route'
          type: array
      type: object
...
    UpdateRouteReq:
      example:
        path: path
        host: host
        load_balancer_strategy: weight-round-robin
        circuit_breaker:
          max_requests: 1
          max_consecutive_failures: 6
          name: name
          interval: 0
          enabled: true
          timeout: 5
        health_check:
          request:
            headers:
              key: headers
            status_code: 2
            method: method
            body: body
          combined_with_circuit_breaker: true
          interval: 5
          type: eth_block_number
          enabled: true
          timeout: 7
      properties:
        circuit_breaker:
          $ref: '#/components/schemas/circuitbreaker.Settings'
        health_check:
          $ref: '#/components/schemas/healthcheck.HealthCheckConfig'
        host:
          description: "Uuid                 string                         `json:\"\
            uuid\" validate:\"required,uuid\"`"
          type: string
        load_balancer_strategy:
          $ref: '#/components/schemas/balancer.LoadBalancerStrategy'
        path:
          type: string
      type: object
...
    healthcheck.HealthCheckConfig:
      example:
        request:
          headers:
            key: headers
          status_code: 2
          method: method
          body: body
        combined_with_circuit_breaker: true
        interval: 5
        type: eth_block_number
        enabled: true
        timeout: 7
      properties:
        combined_with_circuit_breaker:
          type: boolean
        enabled:
          type: boolean
        interval:
          type: integer
        request:
          $ref: '#/components/schemas/healthcheck.Request'
        timeout:
          type: integer
        type:
          $ref: '#/components/schemas/healthcheck.HealthCheckType'
      type: object
    healthcheck.HealthCheckType:
      enum:
      - eth_block_number
      - eth_syncing
      - custom
      type: string
      x-enum-varnames:
      - EthBlockNumberType
      - EthSyncingType
      - CustomType
    healthcheck.Request:
      example:
        headers:
          key: headers
        status_code: 2
        method: method
        body: body
      properties:
        body:
          type: string
        headers:
          additionalProperties:
            type: string
          type: object
        method:
          type: string
        status_code:
          type: integer
      type: object
    plugin.ListPluginsResponse:
      example:
        plugins:
          key:
            names:
            - names
            - names
      properties:
        plugins:
          additionalProperties:
            $ref: '#/components/schemas/plugin.PluginList'
          description: "Map of plugin lists, keyed by an integer ID"
          type: object
      type: object
    plugin.PluginList:
      example:
        names:
        - names
        - names
      properties:
        names:
          description: Repeated field of plugin names
          items:
            type: string
          type: array
      type: object
    plugin.PluginType:
      enum:
      - ResponseInterceptor
      - RequestInterceptor
      - Middleware
      type: string
      x-enum-varnames:
      - PluginTypeResponseInterceptor
      - PluginTypeRequestInterceptor
      - PluginTypeMiddleware
    responses.BadRequestResponse:
      example:
        code: 3
        message: Invalid input parameters
        status: 400
      properties:
        code:
          example: 3
          type: integer
        message:
          example: Invalid input parameters
          type: string
        status:
          example: 400
          type: integer
      type: object
...
    responses.CreatedSuccessResponse-CreateRouteRes:
      example:
        code: 0
        data:
          uuid: uuid
        message: Created Success
        status: 201
      properties:
        code:
          example: 0
          type: integer
        data:
          $ref: '#/components/schemas/CreateRouteRes'
        message:
          example: Created Success
          type: string
        status:
          example: 201
          type: integer
      type: object
...
    responses.DefaultSuccessResponse-Plugins:
      example:
        code: 0
        data:
          RequestInterceptor:
          - name: name
            description: description
            priority: 81
            config: "{}"
          - name: name
            description: description
            priority: 81
            config: "{}"
          Middleware:
          - name: name
            description: description
            priority: 81
            config: "{}"
          - name: name
            description: description
            priority: 81
            config: "{}"
          ResponseInterceptor:
          - name: name
            description: description
            priority: 81
            config: "{}"
          - name: name
            description: description
            priority: 81
            config: "{}"
        message: Created Success
        status: 200
      properties:
        code:
          example: 0
          type: integer
        data:
          $ref: '#/components/schemas/Plugins'
        message:
          example: Created Success
          type: string
        status:
          example: 200
          type: integer
      type: object
    responses.DefaultSuccessResponse-ListRouteRes:
      example:
        code: 0
        data:
          items:
          - path: path
            host: host
            load_balancer_strategy: weight-round-robin
            circuit_breaker:
              max_requests: 1
              max_consecutive_failures: 6
              name: name
              interval: 0
              enabled: true
              timeout: 5
            health_check:
              request:
                headers:
                  key: headers
                status_code: 2
                method: method
                body: body
              combined_with_circuit_breaker: true
              interval: 5
              type: eth_block_number
              enabled: true
              timeout: 7
            uuid: uuid
          - path: path
            host: host
            load_balancer_strategy: weight-round-robin
            circuit_breaker:
              max_requests: 1
              max_consecutive_failures: 6
              name: name
              interval: 0
              enabled: true
              timeout: 5
            health_check:
              request:
                headers:
                  key: headers
                status_code: 2
                method: method
                body: body
              combined_with_circuit_breaker: true
              interval: 5
              type: eth_block_number
              enabled: true
              timeout: 7
            uuid: uuid
        message: Created Success
        status: 200
      properties:
        code:
          example: 0
          type: integer
        data:
          $ref: '#/components/schemas/ListRouteRes'
        message:
          example: Created Success
          type: string
        status:
          example: 200
          type: integer
      type: object
...
    responses.DefaultSuccessResponse-plugin_ListPluginsResponse:
      example:
        code: 0
        data:
          plugins:
            key:
              names:
              - names
              - names
        message: Created Success
        status: 200
      properties:
        code:
          example: 0
          type: integer
        data:
          $ref: '#/components/schemas/plugin.ListPluginsResponse'
        message:
          example: Created Success
          type: string
        status:
          example: 200
          type: integer
      type: object
    responses.DefaultSuccessResponseWithoutData:
      example:
        code: 0
        message: Success
        status: 200
      properties:
        code:
          example: 0
          type: integer
        message:
          example: Success
          type: string
        status:
          example: 200
          type: integer
      type: object
    responses.InternalServerErrorResponse:
      example:
        code: 13
        message: Internal server error
        status: 500
      properties:
        code:
          example: 13
          type: integer
        message:
          example: Internal server error
          type: string
        status:
          example: 500
          type: integer
      type: object
...